changes delete registered team

+++ src/main/java/be/ucll/ip/minor/groep5610/team/domain/TeamService.java
@@ -56,6 +56,9 @@ public class TeamService {
     }
 
     public void deleteTeamById(Long id) {
+        if(getTeam(id).getRegisteredIn().size() > 0){
+            throw new ServiceException(messageSource.getMessage("team.registered.in.regatta", null, LocaleContextHolder.getLocale()));
+        }
         teamRepository.delete(getTeam(id));
     }
